Upgrade dependencies to work with latest PhantomJS
Prebuilt binaries for PhantomJS 1.9 fail on current macOS, High Sierra.  PhantomJS is needed to run tests but is no longer maintained.  This upgrades QUnit dependencies to a version that runs PhantomJS 2.1, which has prebuilt binaries that work on High Sierra.  It is a stopgap solution to enable continuous integration.  Migrating from PhantomJS to headless Chrome would be a better long-term solution, but that is a larger task.