Allow HTML in the popover for variant custom reader
This allows users to configure any arbitrary HTML. I looked in igv-ui's popover component, but this is not using that.

We needed a way to create a custom variant reader and have a button to view more information in the popup dialog. This enables that by doing something like:


` fields.push({ html: '<button>Some button for this variant'});`

as part of the `Variant.prototype.popupData` function in the custom reader.