Allow HTML in the popover for variant custom reader (#1281)
This allows users to configure any arbitrary HTML. I looked in igv-ui's popover component, but this is not using that.

We needed a way to create a custom variant reader and have a button to view more information in the popup dialog. This enables that by doing something like:


` fields.push({ html: '<button>Some button for this variant'});`

as part of the `Variant.prototype.popupData` function in the custom reader.

Former-commit-id: 9095829bb5d7604d111657c16c6473edcfe87d02