Update igv-create.js (#1461)
bug fix - when loading the igv, if you don't provide an explicit `queryParametersSupported: false` in the options object, the igv iterates over current query parameters looking for parameters of certain names (like 'file') and appends them to the config. and if, by any chance, you have corruption in the url (like we had an accidental `&&` in the middle), the igv loading hangs the browser (goes into infinite loop). this simple fix solves this (if a query param is somehow invalid, move on to the next)

Former-commit-id: 657b479ab1a819997f4d1175ba3dea3183b6dfc8